Hawaii fork (#34)
### Description

The `hawaii` hard fork is to make Dogechain from `V1` to `V2`. The major
target is to bring EVM-equivalent back to EVM-compatible, and make full
use of all fundamental infrastructures of Ethereum network.

It is a long run to reach this, and here comes to an end of endless
pains: inefficient network access, high operating node costs, weird
contract errors, easy dust attacks, etc.

Thanks to @0xcb9ff9, @abrahamcruise321 and everyone else who provided
advice and help in making this happen.

### Changes

Notable changes: 

* Correct implementation of data availability, execution layer (EVM),
consensus layer.
* Supports snapshot and caching speeds up the whole network RPC
endpoints.
* Supports snapshot syncing makes future migration easier and cheaper.
* Supports EIP-1559, makes Dogechain back to modern society.
* Migrate Mainnet data which is over 1.3 TB data size.
* Light client is supported, finally.
* More RPC endpoints.
* ...

Well, there are a lot of them, and it’s impossible to describe them all
in one line.
We should update the documentation once the new version released.

---------

Co-authored-by: 0xcb9ff9 <0xcb9ff9@proton.me>